Using bfloat16 instead of float32

+++ hw2/hw2_task.py
@@ -45,7 +45,7 @@ def profile(loop_fn, model, input_ids, trace_name: str):
 
 
 def generate_optimized(optimized_trace_name: str) -> float:
-    model = build_model(torch.float32)
+    model = build_model(torch.bfloat16)
     input_ids = get_input_ids()
     profile(optimized_loop, model, input_ids, optimized_trace_name)
     elapsed = time_generation(optimized_loop, model, input_ids, "Optimized")
@@ -91,9 +91,9 @@ if __name__ == "__main__":
 # ============================================================================
 #
 # Changes made and speedup per fix:
-# 1. KV Caching: Passed `use_cache=True` and `past_key_values` so the model doesn't recompute attention over the entire sequence every step. 
-# 2. bfloat16: Instantiated the model in `torch.bfloat16` instead of `fp32` to drastically improve memory bandwidth and math throughput.
-# 3. torch.inference_mode(): Wrapped the generation loop to prevent PyTorch from building autograd graphs, saving memory and CPU overhead. Speedup 1.02x
+# 1. KV Caching: Passed `use_cache=True` and `past_key_values` so the model doesn't recompute attention over the entire sequence every step. Speedup: 6.28x 
+# 2. bfloat16: Instantiated the model in `torch.bfloat16` instead of `fp32` to drastically improve memory bandwidth and math throughput. 
+# 3. torch.inference_mode(): Wrapped the generation loop to prevent PyTorch from building autograd graphs, saving memory and CPU overhead. Speedup: 1.02x
 # 4. Deferred .item() sync: Removed `.item()` from the hot loop and kept tokens as tensors until the very end, eliminating CPU-GPU synchronization stalls.
 #
 # Biggest impact and why:
